refactored focalplane code for generalized GFA locs
