+++ py/desimodel/focalplane/__init__.py
@@ -1,0 +1,11 @@
+from .geometry import (
+    get_tile_radius_mm, get_tile_radius_deg, get_radius_mm, get_radius_deg,
+    xy2qs, qs2xy,
+    xy2radec, radec2xy,
+    fiber_area_arcsec2,
+    FocalPlane
+)
+
+from .gfa import GFALocations, on_gfa, on_tile_gfa, get_gfa_targets
+
+from .sim import generate_random_vector_field, generate_random_centroid_offsets
